third_party/blink/web_tests/platform/generic/fast/dom/hover-after-dom-delete-expected.txt

Test for http://crbug.com/240722. If the mouse cursor is not visible, no new hover effects should be invoked when the currently hovered node is removed from the DOM. Press any key to delete the red div.

PASS successfullyParsed is true

TEST COMPLETE

Adding the red and blue divs.
PASS blueDiv.offsetTop is 200
PASS redDiv.offsetTop is 50
Mouse is visible, moving it over the red div.
PASS redDiv.innerHTML is "Red hovered."
PASS blueDiv.innerHTML is ""
Mouse is visible, deleting the red div.
PASS blueDiv.offsetTop is 50
PASS blueText is "\"Hovered.\""
Adding the red and blue divs.
PASS blueDiv.offsetTop is 200
PASS redDiv.offsetTop is 50
Mouse is visible, moving it over the red div.
PASS redDiv.innerHTML is "Red hovered."
PASS blueDiv.innerHTML is ""
Setting the mouse cursor to be invisible.
PASS redDiv.innerHTML is "Red hovered."
PASS blueDiv.innerHTML is ""
Mouse is invisible, deleting the red div.
PASS blueDiv.offsetTop is 50
PASS blueText is "none"
